Fix de subservicio

+++ database/migrations/2024_03_30_223744_create_sub_servicios_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('sub_servicios', function (Blueprint $table) {
             $table->id();
             $table->string('nombre', 150);
-            $table->integer('costo')->nullable();
+            $table->decimal('costo', 8,2)->nullable();
             $table->unsignedBigInteger("estatus_id")->nullable();
             $table->foreign("estatus_id")->references("id")->on("estatus")->onDelete('cascade');
             $table->unsignedBigInteger("servicio_id")->nullable();
